requirements.txt
telethon==1.35.0
aiohttp==3.9.1
python-dotenv==1.0.0
